+++ @types/conviva/index.d.ts
@@ -1,2 +1,0 @@
-import './Conviva'
-import './ConvivaExtension'