+++ src/ts/conviva/ConvivaExtension.ts
@@ -1,29 +1,39 @@
 /**
  * The following properties exist in the `window.Conviva` object but they are missing
  * in types of `conviva-core-sdk` package for some reason, so let's add them manually.
+ *
+ * Keep this in the source code, so the types are available in the compiled code therefore in the consuming projects as well.
  */
 
-declare module Conviva {
-  class ContentMetadata {
-    public static readonly StreamType: Conviva.ConvivaConstants['StreamType']
-    public applicationName: string;
+import * as Conviva from '@convivainc/conviva-js-coresdk';
+
+declare module '@convivainc/conviva-js-coresdk' {
+  export class ContentMetadata {
+    public static readonly StreamType: Conviva.ConvivaConstants['StreamType'];
+
+    // Can only be set once
     public assetName: string;
-    public custom: {};
-    public defaultResource: string;
-    public duration: number;
-    public encodedFrameRate: number;
-    public streamType: Conviva.valueof<Conviva.ConvivaConstants['StreamType']>;
-    public streamUrl: string;
+
+    // Can only be set before playback started
     public viewerId: string;
+    public streamType: Conviva.valueof<Conviva.ConvivaConstants['StreamType']>;
+    public applicationName: string;
+    public custom: Record<string, string>;
+    public duration: number;
+
+    // Dynamic
+    public defaultResource: string;
+    public encodedFrameRate: number;
+    public streamUrl: string;
   }
 
-  interface LoggingInterface {
+  export interface LoggingInterface {
     consoleLog(message: string, logLevel: Conviva.valueof<Conviva.ConvivaConstants['LogLevel']>): void;
 
     release(): void;
   }
 
-  class SystemSettings {
+  export class SystemSettings {
     public static readonly LogLevel: Conviva.ConvivaConstants['LogLevel'];
     public allowUncaughtExceptions: boolean;
     public httpTimeout: number;
@@ -31,27 +41,27 @@ declare module Conviva {
     public storageTimeout: number;
   }
 
-  interface TimeInterface {
+  export interface TimeInterface {
     getEpochTimeMs(): number;
 
     release(): void;
   }
 
-  type TimerAction = () => void;
+  export type TimerAction = () => void;
 
-  type TimerCancelFunction = () => void;
+  export type TimerCancelFunction = () => void;
 
-  interface TimerInterface {
+  export interface TimerInterface {
     createTimer(timerAction: TimerAction, intervalMs: number, actionName?: string | null): TimerCancelFunction;
 
     release(): void;
   }
 
-  type HttpRequestCancelFunction = () => void;
+  export type HttpRequestCancelFunction = () => void;
 
-  type HttpRequestCallback = (succeeded: boolean, data: string) => void;
+  export type HttpRequestCallback = (succeeded: boolean, data: string) => void;
 
-  interface HttpInterface {
+  export interface HttpInterface {
     makeRequest(
       httpMethod: 'GET' | 'POST',
       url: string,
@@ -64,11 +74,11 @@ declare module Conviva {
     release(): void;
   }
 
-  type StorageLoadDataCallback = (succeeded: boolean, data: string | null) => void;
+  export type StorageLoadDataCallback = (succeeded: boolean, data: string | null) => void;
 
-  type StorageSaveDataCallback = (succeeded: boolean, data: string | null) => void;
+  export type StorageSaveDataCallback = (succeeded: boolean, data: string | null) => void;
 
-  interface StorageInterface {
+  export interface StorageInterface {
     loadData(storageSpace: string, storageKey: string, callback: StorageLoadDataCallback): void;
 
     saveData(storageSpace: string, storageKey: string, data: string, callback: StorageSaveDataCallback): void;
@@ -76,8 +86,11 @@ declare module Conviva {
     release(): void;
   }
 
-  interface VideoAnalytics {
+  export interface VideoAnalytics {
     reportAppEvent(eventType: string, eventDetail: { [key: string]: string }): void;
     getSessionId(): number;
   }
 }
+
+// This is required, otherwise this file cannot be imported (would fail in runtime without it).
+export {};